% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikGauss.R
\name{LikGauss}
\alias{LikGauss}
\title{Gaussian Likelihood}
\description{
The likelihood of \eqn{y_i} given \eqn{f_i} is normal with mean \eqn{f_i}
and variance \eqn{\sigma_y^2}. It takes a single hyperparameter,
\eqn{\log(\sigma_y)}{log(\sigma_y)}.
}
\section{Warning}{

Note the hyperparameter should be given on the log scale!
}

\section{Note}{

The \code{f_derivative()} method is currently unimplemented for
the Gaussian likelihood
(it is not needed for parameter optimization).
}

\section{Super class}{
\code{\link[gpmss:LikelihoodFunction]{gpmss::LikelihoodFunction}} -> \code{LikGauss}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character vector of length one giving the likelihood
function's name; "Gaussian"}

\item{\code{hypers}}{A numeric vector giving the likelihood function's
hyperparameters}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-lp}{\code{LikGauss$lp()}}
\item \href{#method-f_derivative}{\code{LikGauss$f_derivative()}}
\item \href{#method-new}{\code{LikGauss$new()}}
\item \href{#method-clone}{\code{LikGauss$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lp"></a>}}
\if{latex}{\out{\hypertarget{method-lp}{}}}
\subsection{Method \code{lp()}}{
Compute log probability of outcomes given function values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LikGauss$lp(y, f, hypers = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{The observed outcomes}

\item{\code{f}}{The function values (generally a posterior mean or mode)}

\item{\code{hypers}}{A numeric vector giving hyperparameters for the
likelihood function. If NULL (the default), the hypers data
member is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-f_derivative"></a>}}
\if{latex}{\out{\hypertarget{method-f_derivative}{}}}
\subsection{Method \code{f_derivative()}}{
Compute partial derivatives of log likelihood function with respect
to f
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LikGauss$f_derivative(y, f, hypers = NULL, order = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{The observed outcomes}

\item{\code{f}}{The function values (generally a posterior mean or mode)}

\item{\code{hypers}}{A numeric vector giving hyperparameters for the
likelihood function. If NULL (the default), the hypers data
member is used.}

\item{\code{order}}{An integer vector of length one giving order of
derivative desired; the default is 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new LikGauss object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LikGauss$new(hypers = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hypers}}{A numeric vector giving hyperparameters for the
likelihood function; a numeric vector of length one giving
the (log of) the standard deviation of the normal distribution
for the likelihood}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LikGauss$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
