% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikLogis.R
\name{LikLogis}
\alias{LikLogis}
\title{Logistic Likelihood}
\description{
The likelihood of \eqn{y_i} given \eqn{f_i} is given by the logistic
function with input \eqn{y_i f_i} (where the binary labels for \eqn{y}
are encoded by 1 and -1). It takes no hyperparameters.
}
\details{
The log likelihood is given by

\deqn{
    \sum_i -\log \left(1 + \exp \left( -y_i f_i \right) \right)
}{%
    \sum_i -log ( 1 + exp( -y_i f_i ) )
}
}
\section{Super class}{
\code{\link[gpmss:LikelihoodFunction]{gpmss::LikelihoodFunction}} -> \code{LikLogis}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character vector of length one giving the likelihood
function's name; "logistic"}

\item{\code{hypers}}{A numeric vector giving the likelihood function's
hyperparameters}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-lp}{\code{LikLogis$lp()}}
\item \href{#method-f_derivative}{\code{LikLogis$f_derivative()}}
\item \href{#method-new}{\code{LikLogis$new()}}
\item \href{#method-clone}{\code{LikLogis$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lp"></a>}}
\if{latex}{\out{\hypertarget{method-lp}{}}}
\subsection{Method \code{lp()}}{
Compute log probability of outcomes given function values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LikLogis$lp(y, f, hypers = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{The observed outcomes}

\item{\code{f}}{The function values (generally a posterior mean or mode)}

\item{\code{hypers}}{A numeric vector giving hyperparameters for the
likelihood function. If NULL (the default), the hypers data
member is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-f_derivative"></a>}}
\if{latex}{\out{\hypertarget{method-f_derivative}{}}}
\subsection{Method \code{f_derivative()}}{
Compute partial derivatives of log likelihood function with respect
to f
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LikLogis$f_derivative(y, f, hypers = NULL, order = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{The observed outcomes}

\item{\code{f}}{The function values (generally a posterior mean or mode)}

\item{\code{hypers}}{A numeric vector giving hyperparameters for the
likelihood function. If NULL (the default), the hypers data
member is used.}

\item{\code{order}}{An integer vector of length one giving order of
derivative desired; the default is 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new LikelihoodFunction object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LikLogis$new(hypers = numeric())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hypers}}{A numeric vector giving hyperparameters for the
likelihood function; a numeric vector of length one giving
the (log of) the standard deviation of the normal distribution
for the likelihood}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LikLogis$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
