% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanZero.R
\name{MeanZero}
\alias{MeanZero}
\title{Zero Mean Function}
\description{
A zero mean; for any set of inputs, this function specifies that the
Gaussian process's mean is zero.
}
\details{
As this function specifies a constant prior mean, it has no hyperparameters.
}
\section{Super class}{
\code{\link[gpmss:MeanFunction]{gpmss::MeanFunction}} -> \code{MeanZero}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character vector of length one giving the mean
function's name; "zero"}

\item{\code{hypers}}{A numeric vector giving the mean function's
hyperparameters}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-mean}{\code{MeanZero$mean()}}
\item \href{#method-parameter_derivative}{\code{MeanZero$parameter_derivative()}}
\item \href{#method-input_derivative}{\code{MeanZero$input_derivative()}}
\item \href{#method-new}{\code{MeanZero$new()}}
\item \href{#method-clone}{\code{MeanZero$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mean"></a>}}
\if{latex}{\out{\hypertarget{method-mean}{}}}
\subsection{Method \code{mean()}}{
Compute function prior mean
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanZero$mean(X, hypers = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{The input values (should be a numeric matrix)}

\item{\code{hypers}}{A numeric vector giving hyperparameters for the
mean function. If NULL (the default), the hypers data
member is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parameter_derivative"></a>}}
\if{latex}{\out{\hypertarget{method-parameter_derivative}{}}}
\subsection{Method \code{parameter_derivative()}}{
Compute partial derivatives of mean function with respect to
its hyperparameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanZero$parameter_derivative(X, hypers = NULL, param = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{The input values (should be a numeric matrix)}

\item{\code{hypers}}{A numeric vector giving hyperparameters for the
mean function. If NULL (the default), the hypers data
member is used.}

\item{\code{param}}{An integer vector of length one; which element of
\code{hypers} should the derivative be taken with respect to?
The default is 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-input_derivative"></a>}}
\if{latex}{\out{\hypertarget{method-input_derivative}{}}}
\subsection{Method \code{input_derivative()}}{
Compute partial derivatives of mean function with respect to
its inputs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanZero$input_derivative(X, hypers = NULL, dimension = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{The input values (should be a numeric matrix)}

\item{\code{hypers}}{A numeric vector giving hyperparameters for the
mean function. If NULL (the default), the hypers data
member is used.}

\item{\code{dimension}}{an integer vector of length one giving the dimension
of X with respect to which the derivative is being taken; the
default is 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new MeanZero object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanZero$new(hypers = numeric())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hypers}}{A(n empty) numeric vector giving hyperparameters for
the mean function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanZero$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
