% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanLinear.R
\name{MeanLinear}
\alias{MeanLinear}
\title{Linear Mean Function}
\description{
A linear mean.
}
\details{
This function gives a linear mean; denoting its hyper vector by \eqn{\beta},
for an input matrix \eqn{X}, the prior mean is \eqn{X \beta}.
}
\section{Super class}{
\code{\link[gpmss:MeanFunction]{gpmss::MeanFunction}} -> \code{MeanLinear}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A character vector of length one giving the mean
function's name; "linear"}

\item{\code{hypers}}{A numeric vector giving the mean function's
hyperparameters}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-mean}{\code{MeanLinear$mean()}}
\item \href{#method-parameter_derivative}{\code{MeanLinear$parameter_derivative()}}
\item \href{#method-input_derivative}{\code{MeanLinear$input_derivative()}}
\item \href{#method-new}{\code{MeanLinear$new()}}
\item \href{#method-clone}{\code{MeanLinear$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mean"></a>}}
\if{latex}{\out{\hypertarget{method-mean}{}}}
\subsection{Method \code{mean()}}{
Compute function prior mean
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanLinear$mean(X, hypers = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{The input values (should be a numeric matrix)}

\item{\code{hypers}}{A numeric vector giving hyperparameters for the
mean function. If NULL (the default), the hypers data
member is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parameter_derivative"></a>}}
\if{latex}{\out{\hypertarget{method-parameter_derivative}{}}}
\subsection{Method \code{parameter_derivative()}}{
Compute partial derivatives of mean function with respect to
its hyperparameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanLinear$parameter_derivative(X, hypers = NULL, param = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{The input values (should be a numeric matrix)}

\item{\code{hypers}}{A numeric vector giving hyperparameters for the
mean function. If NULL (the default), the hypers data
member is used.}

\item{\code{param}}{An integer vector of length one; which element of
\code{hypers} should the derivative be taken with respect to?
The default is 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-input_derivative"></a>}}
\if{latex}{\out{\hypertarget{method-input_derivative}{}}}
\subsection{Method \code{input_derivative()}}{
Compute partial derivatives of mean function with respect to
its inputs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanLinear$input_derivative(X, hypers = NULL, dimension = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{The input values (should be a numeric matrix)}

\item{\code{hypers}}{A numeric vector giving hyperparameters for the
mean function. If NULL (the default), the hypers data
member is used.}

\item{\code{dimension}}{an integer vector of length one giving the dimension
of X with respect to which the derivative is being taken; the
default is 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new MeanLinear object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanLinear$new(hypers = numeric())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hypers}}{A numeric vector giving hyperparameters for
the mean function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeanLinear$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
